backend/requirements.txt
fastapi
uvicorn
pyspark==3.5.0
pandas
numpy
psycopg2-binary
pymongo
python-multipart
python-dotenv
chromadb
pypdf
tiktoken
langchain-community
scikit-learn
langchain
langgraph
langchain-openai
sqlalchemy
sentence-transformers
scipy
statsmodels
xgboost
mcp
